Add --json-mini. Use SoM image for /generate endpoint. Update README
Add --json-mini format to make a deki output smaller. It is useful for
LLM to not overuse LLM's context windows. Will be used for AI agents.

Use SoM image for the /generate endpoint. Set of marks prompting is
useful to increase the LLM's image understanding. Read: Set-of-Mark
Prompting Unleashes Extraordinary Visual Grounding in GPT-4V

Add better examples in README